# src/config/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perfume Shop API",
    "description": "RESTful API for a perfume and cologne e-commerce platform with multi-variant product support, allowing flexible volume-based pricing and weight-based sales",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourdomain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/{slug}/activate": {
      "patch": {
        "tags": [
          "Admin - Products"
        ],
        "summary": "Activate a product",
        "description": "Activate an inactive product (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Product activated successfully"
          },
          "400": {
            "description": "Product already active"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{slug}/deactivate": {
      "patch": {
        "tags": [
          "Admin - Products"
        ],
        "summary": "Deactivate a product",
        "description": "Deactivate an active product (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deactivated successfully"
          },
          "400": {
            "description": "Product already inactive"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{slug}/gallery": {
      "post": {
        "tags": [
          "Admin - Products"
        ],
        "summary": "Add gallery images to product",
        "description": "Upload new gallery images (replaces old gallery) (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Gallery updated successfully"
          },
          "400": {
            "description": "No images provided"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Gallery images (max 5 files)"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Products"
        ],
        "summary": "Delete product gallery",
        "description": "Delete all gallery images for a product (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Gallery deleted successfully"
          },
          "400": {
            "description": "Gallery already empty"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{slug}/comments": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product comments",
        "description": "Retrieve all approved comments for a specific product",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/{slug}/comment": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a product comment",
        "description": "Submit a comment and rating for a product",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          }
        ],
        "responses": {
          "201": {
            "description": "Comment submitted successfully"
          },
          "400": {
            "description": "Validation error or already commented"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "Great product! Highly recommended."
                  },
                  "rating": {
                    "type": "integer",
                    "example": 5,
                    "minimum": 1,
                    "maximum": 5
                  }
                },
                "required": [
                  "content",
                  "rating"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    }
  }
}